peer: avoid double TCP connections
When a peer receives its test plan from coordinator with a list of all
peers, it adds TCP connections via set_transport_config().

Instead of connecting to all peers, only connect to those with with
higher IP addresses than ours. That is, the lower IP address peer is the
active side of each TCP connection.