agents: coordinator specifies connection graph
Before this change we'd present an n^2 all-to-all TCP connection graph
to the SDK; each set of peers would establish one connection.

With this change, the coordinator adds a field to the test plan which is
a map from peer id to list of neighbors. The peers now only establish
connections specified for their ID in the plan.